Refactoring some classes. Make SimulationManagement suport SNetS API.
