+++ src/main/java/simulationControl/Main.java
@@ -1,24 +1,21 @@
 package simulationControl;
 
-import java.io.*;
-import java.net.SocketImpl;
-import java.util.*;
-
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseOptions;
 import com.google.firebase.auth.FirebaseCredentials;
 import com.google.firebase.database.*;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-
-import network.Mesh;
-import network.Pair;
-import network.RequestGenerator;
 import simulationControl.distributedProcessing.Client;
 import simulationControl.distributedProcessing.ServerM;
 import simulationControl.distributedProcessing.ServerS;
-import simulationControl.parsers.*;
-import simulator.Simulation;
+import simulationControl.parsers.SimulationConfig;
+import simulationControl.parsers.SimulationRequest;
+import simulationControl.parsers.SimulationServer;
+
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
 
 /**
  * This class has the main method that will instantiate the parsers to read the 
@@ -95,11 +92,10 @@ public class Main {
 
                 if(sr.getStatus().equals("new")) {
                     try {
-                        newRef.child("status").setValue("started");
-                        newRef.child("progress").setValue(0.0);
-                        List<List<Simulation>> allSimulations = createAllSimulations(sr);
+                        newRef.child("status").setValue("started",null);
+                        newRef.child("progress").setValue(0.0,null);
                         //remember to implement with thread
-                        SimulationManagement sm = new SimulationManagement(allSimulations,1);
+                        SimulationManagement sm = new SimulationManagement(sr);
                         sm.startSimulations(new SimulationManagement.SimulationProgressListener() {
                             @Override
                             public void onSimulationProgressUpdate(double progress) {
@@ -110,16 +106,6 @@ public class Main {
                             public void onSimulationFinished() {//do nothing
                             }
                         });
-                        if(sr.getSimulationConfig().getActiveMetrics().BlockingProbability)sr.getResult().blockingProbability = sm.getBlockingProbabilityCsv();
-                        if(sr.getSimulationConfig().getActiveMetrics().BandwidthBlockingProbability)sr.getResult().bandwidthBlockingProbability = sm.getBandwidthBlockingProbabilityCsv();
-                        if(sr.getSimulationConfig().getActiveMetrics().ExternalFragmentation)sr.getResult().externalFragmentation = sm.getExternalFragmentationCsv();
-                        if(sr.getSimulationConfig().getActiveMetrics().RelativeFragmentation)sr.getResult().relativeFragmentation = sm.getRelativeFragmentationCsv();
-                        if(sr.getSimulationConfig().getActiveMetrics().SpectrumUtilization)sr.getResult().spectrumUtilization = sm.getSpectrumUtilizationCsv();
-                        if(sr.getSimulationConfig().getActiveMetrics().TransmittersReceiversRegeneratorsUtilization)sr.getResult().transceiversUtilization = sm.getTransceiversUtilizationCsv();
-                        if(sr.getSimulationConfig().getActiveMetrics().SpectrumSizeStatistics)sr.getResult().spectrumStatistics = sm.getSpectrumStatisticsCsv();
-                        if(sr.getSimulationConfig().getActiveMetrics().EnergyConsumption)sr.getResult().energyConsumption = sm.getEnergyConsumptionCsv();
-                        if(sr.getSimulationConfig().getActiveMetrics().ModulationUtilization)sr.getResult().modulationUtilization = sm.getModulationUtilizationCsv();
-                        if(sr.getSimulationConfig().getActiveMetrics().ConsumedEnergy)sr.getResult().consumedEnergy = sm.getConsumedEnergyCsv();
                         sr.setProgress(1.0);
                         sr.setStatus("finished");
                         newRef.setValue(sr);
@@ -191,14 +177,12 @@ public class Main {
 
         SimulationFileManager sfm = new SimulationFileManager();
         SimulationRequest simulationRequest = sfm.readSimulation(path, name);
-
-        List<List<Simulation>> allSimulations = createAllSimulations(simulationRequest);
         SimulationConfig sc = simulationRequest.getSimulationConfig();
         System.out.println("Threads running: " + sc.getThreads());
         
         //Now start the simulations
         System.out.println("Starting simulations");
-        SimulationManagement sm = new SimulationManagement(allSimulations, sc.getThreads());
+        SimulationManagement sm = new SimulationManagement(simulationRequest);
         
         long start = System.nanoTime();
         
@@ -217,7 +201,6 @@ public class Main {
         long end = System.nanoTime();
         
         System.out.println("saving results");
-        simulationRequest.setResult(sm.getResults());
         sfm.writeSimulation(path,simulationRequest);
         //sm.saveResults(path);
         System.out.println("finish!");
@@ -226,111 +209,5 @@ public class Main {
 		System.out.println("Total simulation time (s): " + (time / 1000000000.0));
     }
 
-    public static SimulationRequest makeSR(String path) throws FileNotFoundException {
-        //Path of the simulation configuration files
-        String separator = System.getProperty("file.separator");
-        String filesPath = path;
-        String networkFilePath = filesPath + separator + "network";
-        String simulationFilePath = filesPath + separator + "simulation";
-        String trafficFilePath = filesPath + separator + "traffic";
-        //String routesFilePath = filesPath + separator + "fixedRoutes";
-        String physicalLayerFilePath = filesPath + separator + "physicalLayer";
-        String othersFilePath = filesPath + separator + "others";
-        Util.projectPath = filesPath;
-
-        //Read files
-        Scanner scanner = new Scanner(new File(networkFilePath));
-        String networkConfigJSON = "";
-        while (scanner.hasNext()) {
-            networkConfigJSON += scanner.next();
-        }
-        scanner = new Scanner(new File(simulationFilePath));
-        String simulationConfigJSON = "";
-        while (scanner.hasNext()) {
-            simulationConfigJSON += scanner.next();
-        }
-        scanner = new Scanner(new File(trafficFilePath));
-        String trafficConfigJSON = "";
-        while (scanner.hasNext()) {
-            trafficConfigJSON += scanner.next();
-        }
-        scanner = new Scanner(new File(physicalLayerFilePath));
-        String physicalLayerConfigJSON = "";
-        while (scanner.hasNext()) {
-            physicalLayerConfigJSON += scanner.next();
-        }
-        scanner = new Scanner(new File(othersFilePath));
-        String othersConfigJSON = "";
-        while (scanner.hasNext()) {
-            othersConfigJSON += scanner.next();
-        }
-
-        Gson gson = new GsonBuilder().create();
-        NetworkConfig nc = gson.fromJson(networkConfigJSON, NetworkConfig.class);
-        SimulationConfig sc = gson.fromJson(simulationConfigJSON, SimulationConfig.class);
-        TrafficConfig tc = gson.fromJson(trafficConfigJSON, TrafficConfig.class);
-        PhysicalLayerConfig plc = gson.fromJson(physicalLayerConfigJSON, PhysicalLayerConfig.class);
-        OthersConfig oc = gson.fromJson(othersConfigJSON,OthersConfig.class);
-        scanner.close();
-
-        SimulationRequest sr = new SimulationRequest();
-        sr.setNetworkConfig(nc);
-        sr.setSimulationConfig(sc);
-        sr.setTrafficConfig(tc);
-        sr.setPhysicalLayerConfig(plc);
-        sr.setOthersConfig(oc);
-        sr.setStatus("new");
-        sr.setProgress(0.0);
-
-        return sr;
-    }
-
-    /**
-     * This method creates all simulations of an experiment
-     *
-     * @throws Exception
-     */
-    public static List<List<Simulation>> createAllSimulations(SimulationRequest sr) throws Exception {
-
-        NetworkConfig nc = sr.getNetworkConfig();
-        SimulationConfig sc = sr.getSimulationConfig();
-        TrafficConfig tc = sr.getTrafficConfig();
-        PhysicalLayerConfig plc = sr.getPhysicalLayerConfig();
-        OthersConfig oc = sr.getOthersConfig();
-        
-        System.out.println("Calculating the modulations transmission ranges");
-        Mesh meshTemp = new Mesh(nc, tc, plc, oc, null);
-        
-        // Create list of simulations
-        List<List<Simulation>> allSimulations = new ArrayList<>(); // Each element of this set is a list with 10 replications from the same load point
-        int i, j;
-        for (i = 0; i < sc.getLoadPoints(); i++) { // Create the simulations for each load point
-            List<Simulation> reps = new ArrayList<>();
-            for (j = 0; j < sc.getReplications(); j++) { // Create the simulations for each replication
-                Mesh m = new Mesh(nc, tc, plc, oc, meshTemp.getModTrDistance());
-                incArrivedRate(m.getPairList(), i);
-                Simulation s = new Simulation(sc, m, i, j);
-                reps.add(s);
-            }
-            allSimulations.add(reps);
-        }
-
-        return allSimulations;
-    }
-
-
-    /**
-     * This method sets the loading point of the simulation in each request generator
-     *
-     * @param pairs Vector<Pair>
-     * @param mult int
-     */
-    private static void incArrivedRate(Vector<Pair> pairs, int mult) {
-        for (Pair pair : pairs) {
-            for (RequestGenerator rg : pair.getRequestGenerators()) {
-                rg.incArrivedRate(mult);
-            }
-        }
-    }
 
 }
